chores: rename utils.ts -> util.ts
